Replace and rename timestamp_create_view.py to timestamp_create.py to fix #16
