+++ common_files/tests/__init__.py
@@ -5,5 +5,5 @@ from common_files.tests.base import BaseAdminTest
 from common_files.tests.start_time import StartTimeListFilterTest
 from common_files.tests.timestamp import TimestampAdminTest
 from common_files.tests.timestamp_active import TimestampActiveAdminTest
-from common_files.tests.timestamp_create_view import TimestampCreateViewTest
+from common_files.tests.timestamp_create import TimestampCreateViewTest
 from common_files.tests.utils import CommonFilesUtilsTest
